testing
background-image: var(--cv-random-wallpaper); is an example of what should be changed in your CSS but i believe it should be done by default. NOTE: it has a default list of images hosted on postimages.org, im not changing that default list because it's what i like, however adding your own list of images, comma separated, will no longer use what's provided. if im not stupid this should already be working