So it works but...
it is personal preference if you want to see people typing or not. Like I don't care about seeing if people are typing in general. Tested in a very spammy discord and saw 20% less cpu usage personally, but constant messages are going out so who really knows how effective blocking it as a whole is